consonants and vowels
Create a JavaScript code that accepts a string of lowercase letters. Print the word followed by how many consonants and vowels it has.
"a", "e", "i", "o", and "u" are considered vowels. 
Every other letter is considered a consonant.
Hint: Define two variables that keep track of the number of consonants and vowels.